Separated installer and launcher functionalities. Updated build.gradle and some documentations regarding this.
