+++ src/main/java/hubturbo/updater/Installer.java
@@ -3,17 +3,10 @@ package hubturbo.updater;
 import address.updater.LibraryDescriptor;
 import address.updater.VersionDescriptor;
 import address.util.*;
-import javafx.application.Application;
 import javafx.application.Platform;
-import javafx.geometry.Insets;
-import javafx.geometry.Pos;
-import javafx.scene.Group;
-import javafx.scene.Scene;
 import javafx.scene.control.Alert;
 import javafx.scene.control.Label;
 import javafx.scene.control.ProgressBar;
-import javafx.scene.layout.VBox;
-import javafx.stage.Stage;
 
 import java.io.File;
 import java.io.IOException;
@@ -27,72 +20,21 @@ import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
 import java.util.Enumeration;
 import java.util.List;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
 import java.util.jar.JarEntry;
 import java.util.jar.JarFile;
 import java.util.stream.Collectors;
 
 /**
- * Unpack all JARs required to run addressbook to the disk.
+ * Unpack all JARs required to runUpdate addressbook onto the disk.
  *
  * Installer JAR will contain all of addressbook required JARs (libraries and dependencies) and
  * the main application JAR.
  */
-public class Installer extends Application {
-    private static final String ERROR_INSTALL = "Failed to install";
-    private static final String ERROR_RUNNING = "Failed to run application";
-    private static final String ERROR_TRY_AGAIN = "Please try again, or contact developer if it keeps failing.";
+public class Installer {
     private static final String LIB_DIR = "lib";
     private static final Path MAIN_APP_FILEPATH = Paths.get(LIB_DIR, new File("resource.jar").getName());
 
-    private final ExecutorService pool = Executors.newSingleThreadExecutor();
-    private ProgressBar progressBar;
-    private Label loadingLabel;
-
-    @Override
-    public void start(Stage primaryStage) throws Exception {
-        showWaitingWindow(primaryStage);
-
-        pool.execute(() -> {
-            try {
-                run();
-            } catch (IOException e) {
-                showErrorDialogAndQuit(ERROR_INSTALL, e.getMessage(), ERROR_TRY_AGAIN);
-            }
-        });
-    }
-
-    private void showWaitingWindow(Stage stage) {
-        stage.setTitle("Initializing");
-        VBox windowMainLayout = new VBox();
-        Group root = new Group();
-        Scene scene = new Scene(root);
-        stage.setScene(scene);
-        scene.setRoot(windowMainLayout);
-
-        loadingLabel = new Label("Initializing. Please wait.");
-
-        progressBar = new ProgressBar(-1.0);
-        progressBar.setPrefWidth(400);
-
-        final VBox vb = new VBox();
-        vb.setSpacing(30);
-        vb.setPadding(new Insets(40));
-        vb.setAlignment(Pos.CENTER);
-        vb.getChildren().addAll(loadingLabel, progressBar);
-        windowMainLayout.getChildren().add(vb);
-
-        stage.show();
-    }
-
-    @Override
-    public void stop() {
-        Platform.exit();
-        System.exit(0);
-    }
-
-    private void run() throws IOException {
+    public void runUpdate(Label label, ProgressBar progressBar) throws IOException {
         try {
             createLibDir();
         } catch (IOException e) {
@@ -111,18 +53,10 @@ public class Installer extends Application {
         }
 
         try {
-            downloadPlatformSpecificComponents();
+            downloadPlatformSpecificComponents(label, progressBar);
         } catch (IOException e) {
             throw new IOException("Failed to download some components.", e);
         }
-
-        try {
-            startMainApplication();
-        } catch (IOException e) {
-            throw new IOException(ERROR_RUNNING, e);
-        }
-
-        stop();
     }
 
     private void createLibDir() throws IOException {
@@ -169,7 +103,7 @@ public class Installer extends Application {
         return FileUtil.getJarFileOfClass(this.getClass()).getName();
     }
 
-    private void downloadPlatformSpecificComponents() throws IOException {
+    private void downloadPlatformSpecificComponents(Label loadingLabel, ProgressBar progressBar) throws IOException {
         System.out.println("Getting platform specific components");
 
         String json = FileUtil.readFromInputStream(Installer.class.getResourceAsStream("/UpdateData.json"));
@@ -207,7 +141,7 @@ public class Installer extends Application {
             Platform.runLater(() -> loadingLabel.setText("Downloading required components. Please wait."));
 
             try {
-                downloadFile(libFile, downloadLink);
+                downloadFile(libFile, downloadLink, progressBar);
             } catch (IOException e) {
                 System.out.println("Failed to download library " + platformDependentLibrary.getFilename());
                 throw e;
@@ -217,19 +151,8 @@ public class Installer extends Application {
         System.out.println("Finished downloading platform-dependent libraries");
     }
 
-    private void startMainApplication() throws IOException {
-        System.out.println("Starting main application");
 
-        String classPath = File.pathSeparator + "lib" + File.separator + "*";
-
-        String command = String.format("java -ea -cp %s address.MainApp", classPath);
-
-        Runtime.getRuntime().exec(command, null, new File(System.getProperty("user.dir")));
-
-        System.out.println("Main application launched");
-    }
-
-    private void downloadFile(File targetFile, URL source) throws IOException {
+    private void downloadFile(File targetFile, URL source, ProgressBar progressBar) throws IOException {
         try (InputStream in = source.openStream()) {
             if (!FileUtil.createFile(targetFile)) {
                 System.out.println("File already exists; file will be replaced");
@@ -244,17 +167,5 @@ public class Installer extends Application {
             System.out.println("Failed to download " + targetFile.toString());
             throw e;
         }
-    }
-
-    private void showErrorDialogAndQuit(String title, String headerText, String contentText) {
-        Platform.runLater(() -> {
-            final Alert alert = new Alert(Alert.AlertType.ERROR);
-            alert.setTitle(title);
-            alert.setHeaderText(headerText);
-            alert.setContentText(contentText);
-
-            alert.showAndWait();
-            stop();
-        });
     }
 }
